chore: move UserAzureLinux3 to utils
